Moved API entities into api package.
