+++ app/src/main/java/com/sokolkatya/myapplication/ui/entities/MovieDetailsItem.kt
@@ -1,17 +1,14 @@
 package com.sokolkatya.myapplication.ui.entities
 
-import com.sokolkatya.myapplication.data.entities.Genre
+import com.sokolkatya.myapplication.data.entities.api.Genre
 
 data class MovieDetailsItem(
-    val adult: Boolean,
     val minimumAge: Int,
     val backdropPath: String,
     val genres: List<Genre>,
     val id: Long,
     val overview: String,
     val posterPath: String,
-    val releaseDate: String,
-    val revenue: Long,
     val runtime: Long,
     val title: String,
     val voteAverage: Double,
